+++ src/main/java/fll/web/playoff/RemoteControlBrackets.java
@@ -18,8 +18,6 @@ import javax.servlet.http.HttpSession;
 import javax.servlet.jsp.PageContext;
 import javax.sql.DataSource;
 
-
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import fll.Team;
@@ -29,11 +27,15 @@ import fll.util.FLLInternalException;
 
 import fll.web.ApplicationAttributes;
 import fll.web.DisplayInfo;
+import fll.web.playoff.BracketData.TopRightCornerStyle;
 
 /**
  * Data for remoteControlBrackets.jsp.
  */
-public class RemoteControlBrackets {
+public final class RemoteControlBrackets {
+
+  private RemoteControlBrackets() {
+  }
 
   private static final org.apache.logging.log4j.Logger LOGGER = org.apache.logging.log4j.LogManager.getLogger();
 
@@ -47,10 +49,7 @@ public class RemoteControlBrackets {
 
     final DataSource datasource = ApplicationAttributes.getDataSource(application);
 
-    // can't close the database connection here as it's used inside
-    // bracketData to create the output, which is called after this scope exits
-    try {
-      final Connection connection = datasource.getConnection();
+    try (Connection connection = datasource.getConnection()) {
 
       pageContext.setAttribute("maxNameLength", Team.MAX_TEAM_NAME_LEN);
 
@@ -64,9 +63,11 @@ public class RemoteControlBrackets {
                                                         4, false, true, h2hBracket.getIndex());
 
         bracketData.addBracketLabels(h2hBracket.getFirstRound());
-        bracketData.addStaticTableLabels();
+        bracketData.addStaticTableLabels(connection);
 
         numRows += bracketData.getNumRows();
+
+        bracketData.generateBracketOutput(connection, TopRightCornerStyle.MEET_TOP_OF_CELL);
 
         allBracketData.add(bracketData);
       }