+++ src/main/java/fll/xml/InvalidEnumCondition.java
@@ -6,27 +6,38 @@
 
 package fll.xml;
 
-import fll.util.FLLRuntimeException;
-
 /**
  * Thrown when an enum condition in a challenge description computed goal is
  * invalid.
  * This typically occurs when a non-enumerated goal is referenced inside an enum
  * condition.
  */
-public class InvalidEnumCondition extends FLLRuntimeException {
+public class InvalidEnumCondition extends ChallengeValidationException {
 
+  /**
+   * Default constructor.
+   */
   public InvalidEnumCondition() {
   }
 
+  /**
+   * @param message {@link #getMessage()}
+   */
   public InvalidEnumCondition(final String message) {
     super(message);
   }
 
+  /**
+   * @param cause {@link #getCause()}
+   */
   public InvalidEnumCondition(final Throwable cause) {
     super(cause);
   }
 
+  /**
+   * @param message {@link #getMessage()}
+   * @param cause {@link #getCause()}
+   */
   public InvalidEnumCondition(final String message,
                               final Throwable cause) {
     super(message, cause);