+++ src/main/java/fll/xml/ui/ConditionStatementEditor.java
@@ -13,8 +13,6 @@ import javax.swing.Box;
 import javax.swing.JComboBox;
 import javax.swing.JPanel;
 
-import org.checkerframework.checker.nullness.qual.Nullable;
-
 import fll.xml.ConditionStatement;
 import fll.xml.GoalScope;
 import fll.xml.InequalityComparison;